pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "Hoveryn"
version = "1.4.0"
description = "Lightweight drone helper code and experiments"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = []

[project.optional-dependencies]
gpio = ["RPi.GPIO>=0.7.0"]
gps = ["gpsd-py3>=0.3.0"]

[project.urls]
Homepage = "https://github.com/qwert1231231/northwind"
Repository = "https://github.com/qwert1231231/northwind.git"
Issues = "https://github.com/qwert1231231/northwind/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["northwind*"]